pedestrian_route_planner/backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.0
requests==2.31.0
# baidu_maps_mcp - 需要手动安装